+++ src/Symfony/Component/DomCrawler/Tests/NativeCrawler/ImageTest.php
@@ -9,48 +9,47 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\DomCrawler\Tests;
+namespace Symfony\Component\DomCrawler\Tests\NativeCrawler;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Component\DomCrawler\Image;
+use Symfony\Component\DomCrawler\NativeCrawler\Image;
 
+/**
+ * @requires PHP 8.4
+ */
 class ImageTest extends TestCase
 {
-    public function testConstructorWithANonImgTag()
+    public function testConstructorWithANonImgTagFromHTMLDocument()
     {
-        $this->expectException(\LogicException::class);
-        $dom = new \DOMDocument();
-        $dom->loadHTML('<html><div><div></html>');
+        $dom = \DOM\HTMLDocument::createFromString('<!DOCTYPE html><html><div><div></html>');
 
+        $this->expectException(\LogicException::class);
         new Image($dom->getElementsByTagName('div')->item(0), 'http://www.example.com/');
     }
 
-    public function testBaseUriIsOptionalWhenImageUrlIsAbsolute()
+    public function testBaseUriIsOptionalWhenImageUrlIsAbsoluteFromHTMLDocument()
     {
-        $dom = new \DOMDocument();
-        $dom->loadHTML('<html><img alt="foo" src="https://example.com/foo" /></html>');
+        $dom = \DOM\HTMLDocument::createFromString('<!DOCTYPE html><html><img alt="foo" src="https://example.com/foo"></html>');
 
         $image = new Image($dom->getElementsByTagName('img')->item(0));
         $this->assertSame('https://example.com/foo', $image->getUri());
     }
 
-    public function testAbsoluteBaseUriIsMandatoryWhenImageUrlIsRelative()
+    public function testAbsoluteBaseUriIsMandatoryWhenImageUrlIsRelativeFromHTMLDocument()
     {
-        $this->expectException(\InvalidArgumentException::class);
-        $dom = new \DOMDocument();
-        $dom->loadHTML('<html><img alt="foo" src="/foo" /></html>');
+        $dom = \DOM\HTMLDocument::createFromString('<!DOCTYPE html><html><img alt="foo" src="/foo"></html>');
 
-        $image = new Image($dom->getElementsByTagName('img')->item(0), 'example.com');
-        $image->getUri();
+        $this->expectException(\InvalidArgumentException::class);
+        new Image($dom->getElementsByTagName('img')->item(0), 'example.com');
     }
 
     /**
      * @dataProvider getGetUriTests
      */
-    public function testGetUri($url, $currentUri, $expected)
+    public function testGetUriFromHTMLDocument($url, $currentUri, $expected)
     {
-        $dom = new \DOMDocument();
-        $dom->loadHTML(\sprintf('<html><img alt="foo" src="%s" /></html>', $url));
+        $dom = \DOM\HTMLDocument::createFromString(sprintf('<!DOCTYPE html><html><img alt="foo" src="%s"></html>', $url));
+
         $image = new Image($dom->getElementsByTagName('img')->item(0), $currentUri);
 
         $this->assertEquals($expected, $image->getUri());